+++ packages/homebridge-ring/security-panel.ts
@@ -1,15 +1,14 @@
-import { BaseDeviceAccessory } from './base-device-accessory'
-import {
+import { BaseDeviceAccessory } from './base-device-accessory.ts'
+import type {
   AlarmMode,
   AlarmState,
-  allAlarmStates,
   RingDevice,
   RingDeviceData,
-  RingDeviceType,
 } from 'ring-client-api'
-import { hap } from './hap'
-import { RingPlatformConfig } from './config'
-import { PlatformAccessory } from 'homebridge'
+import { allAlarmStates, RingDeviceType } from 'ring-client-api'
+import { hap } from './hap.ts'
+import type { RingPlatformConfig } from './config.ts'
+import type { PlatformAccessory } from 'homebridge'
 import { logError, logInfo } from 'ring-client-api/util'
 
 function isValidNightModeBypass(mode?: AlarmMode) {
@@ -17,9 +16,7 @@ function isValidNightModeBypass(mode?: AlarmMode) {
 }
 
 export class SecurityPanel extends BaseDeviceAccessory {
-  private alarmStates: AlarmState[] = this.config.alarmOnEntryDelay
-    ? allAlarmStates
-    : allAlarmStates.filter((x) => x !== 'entry-delay')
+  private alarmStates: AlarmState[]
 
   constructor(
     public readonly device: RingDevice,
@@ -27,6 +24,10 @@ export class SecurityPanel extends BaseDeviceAccessory {
     public readonly config: RingPlatformConfig,
   ) {
     super()
+
+    this.alarmStates = this.config.alarmOnEntryDelay
+      ? allAlarmStates
+      : allAlarmStates.filter((x) => x !== 'entry-delay')
 
     const { Characteristic, Service } = hap,
       validValues = [