+++ packages/homebridge-ring/valve.ts
@@ -5,7 +5,7 @@ import type { RingPlatformConfig } from './config.ts'
 import type { PlatformAccessory } from 'homebridge'
 import { logInfo } from 'ring-client-api/util'
 
-export class UnknownZWaveSwitchSwitch extends BaseDeviceAccessory {
+export class Valve extends BaseDeviceAccessory {
   constructor(
     public readonly device: RingDevice,
     public readonly accessory: PlatformAccessory,
@@ -18,14 +18,23 @@ export class UnknownZWaveSwitchSwitch extends BaseDeviceAccessory {
     this.registerCharacteristic({
       characteristicType: Characteristic.On,
       serviceType: Service.Switch,
-      getValue: (data) => Boolean(data.basicValue),
+      getValue: (data) => this.isOpen(data.valveState),
       setValue: (value) => this.setOnState(value),
     })
   }
 
+  isOpen(status: RingDevice['data']['valveState']): boolean {
+    if (status === 'open') {
+      return true
+    }
+    return false
+  }
+
   setOnState(on: boolean) {
     logInfo(`Turning ${this.device.name} ${on ? 'On' : 'Off'}`)
-
-    return this.device.setInfo({ device: { v1: { basicValue: on ? 255 : 0 } } })
+    if (on) {
+      return this.device.sendCommand('valve.open')
+    }
+    return this.device.sendCommand('valve.close')
   }
 }