+++ packages/ring-client-api/rest-client.ts
@@ -1,4 +1,3 @@
-import got, { Options as RequestOptions, Headers } from 'got'
 import {
   delay,
   fromBase64,
@@ -8,20 +7,32 @@ import {
   logInfo,
   stringify,
   toBase64,
-} from './util'
-import {
+} from './util.ts'
+import type {
   Auth2faResponse,
   AuthTokenResponse,
   SessionResponse,
-} from './ring-types'
+} from './ring-types.ts'
 import { ReplaySubject } from 'rxjs'
 import assert from 'assert'
-import type { Credentials } from '@eneris/push-receiver/dist/types'
+import type { Credentials } from '@eneris/push-receiver/dist/types.d.js'
+import { Agent } from 'undici'
 
-const defaultRequestOptions: RequestOptions = {
+interface RequestOptions extends RequestInit {
+  responseType?: 'json' | 'buffer'
+  timeout?: number
+  json?: object
+  dispatcher?: Agent
+}
+
+const fetchAgent = new Agent({
+    connections: 6,
+    pipelining: 1,
+    keepAliveTimeout: 115000,
+  }),
+  defaultRequestOptions: RequestOptions = {
     responseType: 'json',
     method: 'GET',
-    retry: 0,
     timeout: 20000,
   },
   ringErrorCodes: { [code: number]: string } = {
@@ -34,7 +45,7 @@ const defaultRequestOptions: RequestOptions = {
   clientApiBaseUrl = 'https://api.ring.com/clients_api/',
   deviceApiBaseUrl = 'https://api.ring.com/devices/v1/',
   commandsApiBaseUrl = 'https://api.ring.com/commands/v1/',
-  appApiBaseUrl = 'https://app.ring.com/api/v1/',
+  appApiBaseUrl = 'https://prd-api-us.prd.rings.solutions/api/v1/',
   apiVersion = 11
 
 export function clientApi(path: string) {
@@ -58,43 +69,122 @@ export interface ExtendedResponse {
   timeMillis: number
 }
 
+interface ResponseError extends Error {
+  response: Pick<Response, 'headers' | 'status'> & { body: any }
+}
+
+async function responseToError(response: Response) {
+  const error = new Error() as ResponseError
+  error.response = {
+    headers: response.headers,
+    status: response.status,
+    body: null,
+  }
+
+  try {
+    const bodyText = await response.text()
+
+    try {
+      error.response.body = JSON.parse(bodyText)
+    } catch {
+      error.response.body = bodyText
+    }
+  } catch {
+    // ignore
+  }
+
+  return error
+}
+
 async function requestWithRetry<T>(
   requestOptions: RequestOptions & { url: string; allowNoResponse?: boolean },
+  retryCount = 0,
 ): Promise<T & ExtendedResponse> {
+  if (typeof fetch !== 'function') {
+    throw new Error(
+      `Your current NodeJS version (${process.version}) is too old to support this plugin.  Please upgrade to the latest LTS version of NodeJS.`,
+    )
+  }
+
   try {
-    const options = {
-        ...defaultRequestOptions,
-        ...requestOptions,
-      },
-      { headers, body } = (await got(options)) as {
-        headers: Headers
-        body: any
-      },
-      data = body as T & ExtendedResponse
-    if (data !== null && typeof data === 'object') {
-      if (headers.date) {
-        data.responseTimestamp = new Date(headers.date as string).getTime()
+    if (requestOptions.json || requestOptions.responseType === 'json') {
+      requestOptions.headers = {
+        ...requestOptions.headers,
+        'Content-Type': 'application/json',
+        Accept: 'application/json',
       }
 
-      if (headers['x-time-millis']) {
-        data.timeMillis = Number(headers['x-time-millis'])
+      if (requestOptions.json) {
+        requestOptions.body = JSON.stringify(requestOptions.json)
+      }
+      delete requestOptions.json
+    }
+
+    const options = {
+      ...defaultRequestOptions,
+      ...requestOptions,
+      dispatcher: fetchAgent,
+    }
+
+    // If a timeout is provided, create an AbortSignal for it
+    if (options.timeout && !options.signal) {
+      options.signal = AbortSignal.timeout(options.timeout)
+    }
+
+    // make the fetch request
+    const response = await fetch(options.url, options),
+      headers = response.headers
+
+    if (!response.ok) {
+      const error = await responseToError(response)
+      throw error
+    }
+
+    let data: T & ExtendedResponse
+
+    if (options.responseType === 'buffer') {
+      const arrayBuffer = await response.arrayBuffer()
+      data = Buffer.from(arrayBuffer) as any
+    } else {
+      const text = await response.text()
+      try {
+        data = JSON.parse(text)
+      } catch {
+        data = text as any
+      }
+    }
+
+    if (data !== null && typeof data === 'object') {
+      const date = headers.get('date')
+      if (date) {
+        data.responseTimestamp = new Date(date).getTime()
+      }
+
+      const xTime = headers.get('x-time-millis')
+      if (xTime) {
+        data.timeMillis = Number(xTime)
       }
     }
     return data
   } catch (e: any) {
     if (!e.response && !requestOptions.allowNoResponse) {
-      logError(
-        `Failed to reach Ring server at ${requestOptions.url}.  ${e.message}.  Trying again in 5 seconds...`,
-      )
-      if (e.message.includes('NGHTTP2_ENHANCE_YOUR_CALM')) {
+      if (retryCount > 0) {
+        let detailedError = `Error: ${e.message}`
+        detailedError += e.cause?.message ? `, Cause: ${e.cause.message}` : ''
+        detailedError += e.cause?.code ? `, Code: ${e.cause.code}` : ''
         logError(
-          `There is a known issue with your current NodeJS version (${process.version}).  Please see https://github.com/dgreif/ring/wiki/NGHTTP2_ENHANCE_YOUR_CALM-Error for details`,
+          `Retry #${retryCount} failed to reach Ring server at ${requestOptions.url}.  ${detailedError}.  Trying again in 5 seconds...`,
         )
+        if (e.message.includes('NGHTTP2_ENHANCE_YOUR_CALM')) {
+          logError(
+            `There is a known issue with your current NodeJS version (${process.version}).  Please see https://github.com/dgreif/ring/wiki/NGHTTP2_ENHANCE_YOUR_CALM-Error for details`,
+          )
+        }
+        logDebug(e)
       }
-      logDebug(e)
 
       await delay(5000)
-      return requestWithRetry(requestOptions)
+      return requestWithRetry(requestOptions, retryCount + 1)
     }
     throw e
   }
@@ -136,7 +226,7 @@ function parseAuthConfig(rawRefreshToken?: string): AuthConfig | undefined {
     assert(config)
     assert(config.rt)
     return config
-  } catch (_) {
+  } catch {
     return {
       rt: rawRefreshToken,
     }
@@ -144,13 +234,9 @@ function parseAuthConfig(rawRefreshToken?: string): AuthConfig | undefined {
 }
 
 export class RingRestClient {
-  public refreshToken =
-    'refreshToken' in this.authOptions
-      ? this.authOptions.refreshToken
-      : undefined
-  private authConfig = parseAuthConfig(this.refreshToken)
-  private hardwareIdPromise =
-    this.authConfig?.hid || getHardwareId(this.authOptions.systemId)
+  public refreshToken
+  private authConfig
+  private hardwareIdPromise
   private _authPromise: Promise<AuthTokenResponse> | undefined
   private timeouts: ReturnType<typeof setTimeout>[] = []
   private clearPreviousAuth() {
@@ -189,15 +275,21 @@ export class RingRestClient {
     newRefreshToken: string
   }>(1)
   public onSession = new ReplaySubject<SessionResponse>(1)
-  public readonly baseSessionMetadata = {
-    api_version: apiVersion,
-    device_model:
-      this.authOptions.controlCenterDisplayName ?? 'ring-client-api',
-  }
+  public readonly baseSessionMetadata
 
   constructor(
     private authOptions: (EmailAuth | RefreshTokenAuth) & SessionOptions,
-  ) {}
+  ) {
+    this.refreshToken =
+      'refreshToken' in authOptions ? authOptions.refreshToken : undefined
+    this.authConfig = parseAuthConfig(this.refreshToken)
+    this.hardwareIdPromise =
+      this.authConfig?.hid || getHardwareId(authOptions.systemId)
+    this.baseSessionMetadata = {
+      api_version: apiVersion,
+      device_model: authOptions.controlCenterDisplayName ?? 'ring-client-api',
+    }
+  }
 
   private getGrantData(twoFactorAuthCode?: string) {
     if (this.authConfig?.rt && !twoFactorAuthCode) {
@@ -279,13 +371,13 @@ export class RingRestClient {
             : ''
 
       if (
-        response.statusCode === 412 || // need 2fa code
-        (response.statusCode === 400 &&
+        response.status === 412 || // need 2fa code
+        (response.status === 400 &&
           responseError.startsWith('Verification Code')) // invalid 2fa code entered
       ) {
         this.using2fa = true
 
-        if (response.statusCode === 400) {
+        if (response.status === 400) {
           this.promptFor2fa = 'Invalid 2fa code entered.  Please try again.'
           throw new Error(responseError)
         }
@@ -349,15 +441,15 @@ export class RingRestClient {
         this.onSession.next(session)
         return session
       } catch (e: any) {
-        const response = e.response || {}
+        const response = (e as ResponseError).response || {}
 
-        if (response.statusCode === 401) {
+        if (response.status === 401) {
           await this.refreshAuth()
           return this.getSession()
         }
 
-        if (response.statusCode === 429) {
-          const retryAfter = e.response.headers['retry-after'],
+        if (response.status === 429) {
+          const retryAfter = e.response.headers.get('retry-after'),
             waitSeconds = isNaN(retryAfter)
               ? 200
               : Number.parseInt(retryAfter, 10)
@@ -420,21 +512,21 @@ export class RingRestClient {
         },
       })
     } catch (e: any) {
-      const response = e.response || {}
+      const response = (e as ResponseError).response || {}
 
-      if (response.statusCode === 401) {
+      if (response.status === 401) {
         await this.refreshAuth()
         return this.request(options)
       }
 
-      if (response.statusCode === 504) {
+      if (response.status === 504) {
         // Gateway Timeout.  These should be recoverable, but wait a few seconds just to be on the safe side
         await delay(5000)
         return this.request(options)
       }
 
       if (
-        response.statusCode === 404 &&
+        response.status === 404 &&
         response.body &&
         Array.isArray(response.body.errors)
       ) {
@@ -459,7 +551,7 @@ export class RingRestClient {
         )
       }
 
-      if (response.statusCode === 404 && url.startsWith(clientApiBaseUrl)) {
+      if (response.status === 404 && url.startsWith(clientApiBaseUrl)) {
         logError('404 from endpoint ' + url)
         if (response.body?.error?.includes(hardwareId)) {
           logError(
@@ -474,10 +566,10 @@ export class RingRestClient {
         throw new Error('Not found with response: ' + stringify(response.body))
       }
 
-      if (response.statusCode) {
+      if (response.status) {
         logError(
           `Request to ${url} failed with status ${
-            response.statusCode
+            response.status
           }. Response body: ${stringify(response.body)}`,
         )
       } else if (!options.allowNoResponse) {