+++ packages/ring-client-api/ring-intercom.ts
@@ -1,13 +1,14 @@
-import {
+import type {
   IntercomHandsetAudioData,
   PushNotification,
-  PushNotificationAction,
-} from './ring-types'
-import { clientApi, commandsApi, RingRestClient } from './rest-client'
+} from './ring-types.ts'
+import { PushNotificationAction } from './ring-types.ts'
+import type { RingRestClient } from './rest-client.ts'
+import { clientApi, commandsApi } from './rest-client.ts'
 import { BehaviorSubject, Subject } from 'rxjs'
 import { distinctUntilChanged, map } from 'rxjs/operators'
-import { getBatteryLevel } from './ring-camera'
-import { logError } from './util'
+import { getBatteryLevel } from './ring-camera.ts'
+import { logError } from './util.ts'
 
 export class RingIntercom {
   id
@@ -104,9 +105,16 @@ export class RingIntercom {
   }
 
   processPushNotification(notification: PushNotification) {
-    if (notification.action === PushNotificationAction.Ding) {
+    if (
+      'android_config' in notification &&
+      notification.android_config.category ===
+        PushNotificationAction.IntercomDing
+    ) {
       this.onDing.next()
-    } else if (notification.action === PushNotificationAction.IntercomUnlock) {
+    } else if (
+      'gcmData' in notification.data &&
+      notification.data.gcmData.action === PushNotificationAction.IntercomUnlock
+    ) {
       this.onUnlocked.next()
     }
   }